feat(leetcode): remove duplicates from sorted array
Use a slow pointer to "lock" the "wanted" element, and use a fast pointer to move forward along the list and look for new unique elements in the list.
Or, in other words, the current slow pointer is used to locate the latest unique number for the results, and fast is used for iterating and discovery.

Have fast advanced in every iteration, but slow is only advanced when two pointers are onto two different elements.

That means, the elements after nums[slow] and before nums[fast] are numbers we've seen before and don't need anymore (one copy of these numbers is already saved before the current slow (inclusive)).

Therefore, in order to have this newly discovered (unseen) number pointed by the current fast to the front of the array for the final answer, we just need to swap this newly discovered number to the location that follows the current slow pointer, with one of the seen numbers (we don't need it for the answer regardlessly), and then advance the slow in the same iteration to lock this new number.

Related issue: #1